requirements.txt
aiohttp
python-telegram-bot
aioresponses
pandas
xlsxwriter
packaging